% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConversion-method.R
\name{getConversion,CosiaAnnotate-method}
\alias{getConversion,CosiaAnnotate-method}
\title{Converting Gene Identifiers between the same species and different species}
\usage{
\S4method{getConversion}{CosiaAnnotate}(object)
}
\arguments{
\item{object}{CosiaAnnotate.}
}
\value{
A dataframe with input identifiers and matching gene identifier conversion of the output species
}
\description{
Converting Gene Identifiers between the same species and different species
}
\examples{
getConversion(mouse_data)
getConversion(rat_data)
}
