% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getConversion-method.R
\name{getConversion,CoSIAn-method}
\alias{getConversion,CoSIAn-method}
\title{getConversion Method}
\usage{
\S4method{getConversion}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
CoSIAn object with converted_id slot filled
}
\description{
getConversion Method
}
\examples{
Kidney_gene_conversion<-CoSIA::getConversion(Kidney_Genes)
}
