% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getConversion-method.R
\name{getConversion}
\alias{getConversion}
\title{getConversion Generic}
\usage{
getConversion(object)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled}
}
\description{
getConversion Generic
}
