% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getConversions-method.R
\name{getConversions,CoSIAn-method}
\alias{getConversions,CoSIAn-method}
\title{getConversions Method}
\usage{
\S4method{getConversions}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
CoSIAn object with converted_id slot filled
}
\description{
getConversions Method
}
\examples{
Kidney_gene_conversion<-CoSIA::getConversions(Kidney_Genes)
}
