% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getGEx-method.R
\name{getGEx,CoSIAn-method}
\alias{getGEx,CoSIAn-method}
\title{getGEx Method}
\usage{
\S4method{getGEx}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
CoSIAn Object with gex slot filled
}
\description{
getGEx Method
}
\examples{
Kidney_gene_gex<- getGEx(Kidney_gene_conversion)
}
