% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getGEx-method.R
\name{getGEx}
\alias{getGEx}
\title{getGEx Generic}
\usage{
getGEx(object)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled with converted_id slot filled}
}
\description{
getGEx Generic
}
