% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getGExMetrics-method.R
\name{getGExMetrics,CoSIAn-method}
\alias{getGExMetrics,CoSIAn-method}
\title{getGExMetrics Method}
\usage{
\S4method{getGExMetrics}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
CoSIAn Object with metric slot filled
}
\description{
getGExMetrics Method
}
\examples{
Kidney_gene_metric<-getGExMetrics(Kidney_gene_gex)
}
