% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-getGExMetrics-method.R
\name{getGExMetrics}
\alias{getGExMetrics}
\title{getGExMetrics Generic}
\usage{
getGExMetrics(object)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled with converted_id slot filled}
}
\description{
getGExMetrics Generic
}
