% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-utility-method.R
\name{getTissues}
\alias{getTissues}
\title{getTissues}
\usage{
getTissues(species)
}
\arguments{
\item{species}{name of a species or multiple species that you want to get 
available tissue list for}
}
\value{
list of tissues that are common/available among the species or 
multiple species inputted
}
\description{
getTissues
}
\examples{
tissue<-getTissues(c('m_musculus'))
}
