% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_id_dataframe.R
\name{missing_id_column}
\alias{missing_id_column}
\title{Missing ID Column}
\usage{
missing_id_column(input_data, output_data)
}
\arguments{
\item{input_data}{The originally character vector of column that contains the gene identifiers}

\item{output_data}{The annotated character vector or column that contains the same gene identifiers that were originally inputted}
}
\value{
a data frame that contains TRUE/FALSE for whether the gene identifier was mapped
}
\description{
Missing ID Column
}
