% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotCVGEx,CoSIAn-method}
\alias{plotCVGEx,CoSIAn-method}
\title{plotCVGEx Method}
\usage{
\S4method{plotCVGEx}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\description{
plotCVGEx Method
}
