% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotDSGEx}
\alias{plotDSGEx}
\title{plotDSGEx Generic}
\usage{
plotDSGEx(object)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled in as well as the converted_id and metric slot filled}
}
\description{
plotDSGEx Generic
}
