% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotSpeciesGEx,CoSIAn-method}
\alias{plotSpeciesGEx,CoSIAn-method}
\title{plotSpeciesGEx Method}
\usage{
\S4method{plotSpeciesGEx}{CoSIAn}(object, single_tissue, single_gene)
}
\arguments{
\item{object}{CoSIAn.}
}
\description{
plotSpeciesGEx Method
}
