% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotSpeciesGEx}
\alias{plotSpeciesGEx}
\title{plotSpeciesGEx Generic}
\usage{
plotSpeciesGEx(object, single_tissue, single_gene)
}
\arguments{
\item{object}{}
}
\description{
plotSpeciesGEx Generic
}
