% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotTissueGEx}
\alias{plotTissueGEx}
\title{plotTissueGEx Generic}
\usage{
plotTissueGEx(object, single_species, single_gene)
}
\arguments{
\item{object}{}
}
\description{
plotTissueGEx Generic
}
