% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-utility-method.R
\name{viewCoSIAn,CoSIAn-method}
\alias{viewCoSIAn,CoSIAn-method}
\title{viewCoSIAn}
\usage{
\S4method{viewCoSIAn}{CoSIAn}(object, slot_name)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled}
}
\value{
slots in CoSIAn object
}
\description{
viewCoSIAn
}
\examples{
Kidney_Genes <- CoSIAn(
    gene_set = c("ENSG00000008710", "ENSG00000152217"),
    i_species = "h_sapiens", input_id = "Ensembl_id",
    o_species = c(
        "d_melanogaster"
    ),
    output_ids = c("Ensembl_id", "Symbol"), mapping_tool = "annotationDBI",
    ortholog_database = "HomoloGene", map_tissues = "heart",
    map_species = c("d_melanogaster"), metric_type = "DS_Gene"
)
viewCoSIAn(Kidney_Genes, converted_id)
}
