% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-utility-method.R
\name{viewCoSIAn}
\alias{viewCoSIAn}
\title{viewCoSIAn Generics}
\usage{
viewCoSIAn(object, ...)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled}
}
\value{
initializes a generic function for viewCoSIAn as preparation for
defining the viewCoSIAn Method
}
\description{
viewCoSIAn Generics
}
\examples{
Kidney_Genes <- CoSIAn(
    gene_set = c("ENSG00000008710", "ENSG00000152217"),
    i_species = "h_sapiens", input_id = "Ensembl_id",
    o_species = c(
        "d_melanogaster"
    ),
    output_ids = c("Ensembl_id", "Symbol"), mapping_tool = "annotationDBI",
    ortholog_database = "HomoloGene", map_tissues = "heart",
    map_species = c("d_melanogaster"), metric_type = "DS_Gene"
)
viewCoSIAn(Kidney_Genes)
}
