% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-class.R
\docType{class}
\name{CoSIAn}
\alias{CoSIAn}
\title{CoSIAn Class}
\usage{
CoSIAn(
  gene_set,
  i_species,
  input_id,
  o_species,
  output_ids,
  mapping_tool = "annotationDBI",
  ortholog_database = "HomoloGene",
  map_tissues,
  map_species,
  metric_type
)
}
\arguments{
\item{gene_set}{A singular gene or a list of genes.}

\item{i_species}{The species corresponding to gene_set.}

\item{input_id}{The type of id corresponding to gene_set.}

\item{o_species}{The singular or list of species that the gene set is being converted to.}

\item{output_ids}{The singular or list of id types that the gene set is being converted to.}

\item{mapping_tool}{The mapping tool, BiomaRt or annotationDBI, being used to map the conversion between IDs.}

\item{ortholog_database}{The ortholog database, HomoloGene or NCBIOrtho, being used to map the conversion between species.}

\item{map_tissues}{A list of tissues being investigated}

\item{map_species}{A list of species being investigated}

\item{metric_type}{A list of possible metric the user wants to calculate.}
}
\value{
An S4 \code{CoSIAn} object with character vector(s) as slots.
}
\description{
The \code{CoSIAn} constructor creates a \code{CoSIAn} object from character vector(s).
}
\section{Slots}{

\describe{
\item{\code{gene_set}}{character. A singular gene or a list of genes.}

\item{\code{i_species}}{character. The species corresponding to gene_set.}

\item{\code{input_id}}{character. The type of id corresponding to gene_set.}

\item{\code{o_species}}{character. The singular or list of species that the gene set is being converted to.}

\item{\code{output_ids}}{character. The singular or list of id types that the gene set is being converted to.}

\item{\code{mapping_tool}}{character. The mapping tool, BiomaRt or annotationDBI, being used to map the conversion between IDs.}

\item{\code{ortholog_database}}{character. The ortholog database, HomoloGene or NCBIOrtho, being used to map the conversion between species.}

\item{\code{converted_id}}{data frame. Output of getConversion.}

\item{\code{map_tissues}}{character. A list of tissues being investigated.}

\item{\code{map_species}}{character. A list of species being investigated.}

\item{\code{gex}}{data frame. Output of gene expression data.}

\item{\code{metric_type}}{character. A list of possible metric the user wants to calculate.}

\item{\code{metric}}{data frame. Output of coefficient of variation of gene expression data.}
}}

\examples{
Kidney_Genes<-CoSIAn(gene_set = c("ENSG00000008710","ENSG00000118762","ENSG00000152217"),
i_species = "h_sapiens",input_id = "Ensembl_id",o_species = c("d_melanogaster","m_musculus",
"h_sapiens", "d_rerio","c_elegans","r_norvegicus"),output_ids = c("Ensembl_id","Symbol"), 
mapping_tool = "annotationDBI",ortholog_database = "HomoloGene",map_tissues = "heart", 
map_species = c("m_musculus"),metric_type = "DS_Gene")
}
