% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotDSGEx,CoSIAn-method}
\alias{plotDSGEx,CoSIAn-method}
\title{plotDSGEx Method}
\usage{
\S4method{plotDSGEx}{CoSIAn}(object)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
plot object
}
\description{
plotDSGEx Method
}
\examples{
plotDSGEx(Kidney_gene_metric)
}
