% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotSpeciesGEx,CoSIAn-method}
\alias{plotSpeciesGEx,CoSIAn-method}
\title{plotSpeciesGEx Method}
\usage{
\S4method{plotSpeciesGEx}{CoSIAn}(object, single_tissue, single_gene)
}
\arguments{
\item{object}{CoSIAn object with all user accessible slots filled in as well as the converted_id and gex slot filled}

\item{single_tissue}{one tissue that the user wants to investigate across the mapped species}

\item{single_gene}{one ensembl id that the user wants to investigate across the mapped species}
}
\value{
plot object
}
\description{
plotSpeciesGEx Method
}
\examples{
Kidney_Genes<-CoSIAn(gene_set = c('ENSG00000008710','ENSG00000118762','ENSG00000152217'),
i_species = 'h_sapiens',input_id = 'Ensembl_id',o_species = c('d_melanogaster','m_musculus',
'h_sapiens', 'd_rerio','c_elegans','r_norvegicus'),output_ids = c('Ensembl_id','Symbol'), 
mapping_tool = 'annotationDBI',ortholog_database = 'HomoloGene',map_tissues = 'heart', 
map_species = c('m_musculus'),metric_type = 'DS_Gene')
Kidney_gene_conversion<-CoSIA::getConversion(Kidney_Genes)
load('~/Desktop/EH_Data.RData')
Kidney_gene_gex<- getGEx(Kidney_gene_conversion)
plotSpeciesGEx(Kidney_gene_gex,'liver','ENSG00000008710')
plotSpeciesGEx(Kidney_gene_gex,'brain','ENSG00000118762')
}
