% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoSIA-plots-method.R
\name{plotTissueGEx,CoSIAn-method}
\alias{plotTissueGEx,CoSIAn-method}
\title{plotTissueGEx Method}
\usage{
\S4method{plotTissueGEx}{CoSIAn}(object, single_species, single_gene)
}
\arguments{
\item{object}{CoSIAn.}
}
\value{
plot object
}
\description{
plotTissueGEx Method
}
\examples{
plotTissueGEx(Kidney_gene_gex,"h_sapiens","ENSG00000008710")
plotTissueGEx(Kidney_gene_gex,"m_musculus","ENSG00000008710")
}
